requirements.txt
# Data
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# MLOps
mlflow==2.10.2
dvc==3.40.0

# API
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# Monitoring
prometheus-client==0.20.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1

# Testing
pytest==8.0.0
httpx==0.27.0